+++ nextcloudappstore/core/api/v1/tests/__init__.py
@@ -4,4 +4,3 @@ from .test_parser import *
 from .test_release_importer import *
 from .test_app_release_provider import *
 from .test_app_register import *
-from .test_deletion_log import *
