+++ nextcloudappstore/core/caching.py
@@ -1,7 +1,7 @@
 from typing import List, Tuple, Any
 from django.db.models import Max, QuerySet
-from nextcloudappstore.core.api.v1.models import AppReleaseDeleteLog
-from nextcloudappstore.core.models import App, AppRelease, Category, AppRating
+from nextcloudappstore.core.models import App, AppReleaseDeleteLog, Category, \
+    AppRating
 
 
 def create_etag(pairs: List[Tuple[QuerySet, str]]) -> str:
@@ -16,16 +16,24 @@ def create_etag(pairs: List[Tuple[QuerySet, str]]) -> str:
     return str(max(result, default=''))
 
 
-def app_etag(request: Any, version: str) -> str:
+def apps_etag(request: Any, version: str) -> str:
     return create_etag([
         (App.objects.all(), 'last_release'),
         (AppReleaseDeleteLog.objects.all(), 'last_modified'),
     ])
 
 
-def category_etag(request: Any) -> str:
+def app_etag(request: Any, id: str) -> str:
+    return str(App.objects.get(id=id).last_modified) or ''
+
+
+def app_rating_etag(request: Any, id: str) -> str:
+    return create_etag([(AppRating.objects.filter(app__id=id), 'rated_at')])
+
+
+def categories_etag(request: Any) -> str:
     return create_etag([(Category.objects.all(), 'last_modified')])
 
 
-def app_rating_etag(request: Any) -> str:
+def app_ratings_etag(request: Any) -> str:
     return create_etag([(AppRating.objects.all(), 'rated_at')])
