+++ nextcloudappstore/core/tests/__init__.py
@@ -2,3 +2,4 @@ from .test_versioning import *
 from .test_rating import *
 from .test_compatibility import *
 from .test_search import *
+from .test_deletion_log import *
