+++ nextcloudappstore/core/tests/test_deletion_log.py
@@ -1,12 +1,9 @@
 from django.contrib.auth import get_user_model
 from django.test import TestCase
-
-from nextcloudappstore.core.api.v1.models import AppReleaseDeleteLog
-from nextcloudappstore.core.models import App, AppRelease
+from nextcloudappstore.core.models import App, AppRelease, AppReleaseDeleteLog
 
 
 class DeletionLogTest(TestCase):
-
     def test_delete_app(self):
         user = get_user_model().objects.create(username='john')
         app = App.objects.create(owner=user, id='news')
